muviassignment/server/requirements.txt
Django==1.11.8
djangorestframework==3.7.3
mysqlclient==1.3.7
